requirements.txt
asgiref==3.5.0
astroid==2.3.3
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.1.3
cloudinary==1.25.0
colorama==0.4.5
cssbeautifier==1.14.6
dj-database-url==0.5.0
Django==3.2
django-cloudinary-storage==0.3.0
django-cockroachdb==3.2
django-crispy-forms==1.9.0
django-endless-pagination==2.0
django-heroku==0.3.1
django-hitcount==1.3.1
django-mdeditor==0.1.20
django-summernote==0.8.11.6
django-taggit==2.1.0
django-watson==1.5.4
djlint==1.15.0
EditorConfig==0.12.3
gunicorn==20.0.4
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.3
importlib-metadata==4.12.0
isort==4.3.21
jsbeautifier==1.14.6
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.1.1
pathspec==0.10.1
Pillow==9.0.0
psycopg2==2.9.3
pylint==2.4.4
python-decouple==3.1
pytz==2019.3
PyYAML==6.0
regex==2022.9.13
requests==2.27.1
six==1.14.0
sqlparse==0.3.1
tomli==2.0.1
tqdm==4.64.1
urllib3==1.26.8
uWSGI==2.0.20
virtualenv==16.7.9
whitenoise==5.0.1
wrapt==1.11.2
zipp==3.8.1
